# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Interview Backend API",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:18080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/job-seeker/jobs/{jobId}/applications": {
      "post": {
        "tags": [
          "job-seeker-application-controller"
        ],
        "operationId": "apply",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobApplicationCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseJobApplicationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/job-seeker/applications/{applicationId}/withdraw": {
      "post": {
        "tags": [
          "job-seeker-application-controller"
        ],
        "operationId": "withdraw",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseJobApplicationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/public/jobs": {
      "get": {
        "tags": [
          "public-job-controller"
        ],
        "operationId": "findPublicJobs",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "location",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "skillIds",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int64"
              }
            }
          },
          {
            "name": "workMode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "jobType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePagePublicJobResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/public/jobs/{jobId}": {
      "get": {
        "tags": [
          "public-job-controller"
        ],
        "operationId": "getPublicJob",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePublicJobResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/job-seeker/applications/{applicationId}": {
      "get": {
        "tags": [
          "job-seeker-application-controller"
        ],
        "operationId": "getMyApplication",
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseJobApplicationResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "JobPostSectionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "sectionType": {
            "type": "string",
            "enum": [
              "DESCRIPTION",
              "REQUIREMENT_RESPONSIBILITY",
              "BENEFIT",
              "QUALIFICATION",
              "NICE_TO_HAVE",
              "ABOUT_ROLE"
            ]
          },
          "title": {
            "type": "string"
          },
          "contentMarkdown": {
            "type": "string"
          },
          "contentText": {
            "type": "string"
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "JobPostSkillResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "skillId": {
            "type": "integer",
            "format": "int64"
          },
          "skillName": {
            "type": "string"
          },
          "skillType": {
            "type": "string"
          },
          "requiredLevel": {
            "type": "string"
          }
        }
      },
...
      "JobApplicationCreateRequest": {
        "type": "object",
        "properties": {
          "resumeId": {
            "type": "integer",
            "format": "int64"
          },
          "coverLetter": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 0
          }
        }
      },
      "ApiResponseJobApplicationResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/JobApplicationResponse"
          }
        }
      },
      "JobApplicationResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "jobId": {
            "type": "integer",
            "format": "int64"
          },
          "jobTitle": {
            "type": "string"
          },
          "resumeId": {
            "type": "integer",
            "format": "int64"
          },
          "resumeTitle": {
            "type": "string"
          },
          "coverLetter": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "SUBMITTED",
              "UNDER_REVIEW",
              "AI_INTERVIEW_REQUIRED",
              "AI_INTERVIEW_IN_PROGRESS",
              "AI_INTERVIEW_FAILED",
              "MODERATOR_REVIEW_PENDING",
              "AI_INTERVIEW_PASSED",
              "SHORTLISTED",
              "HUMAN_INTERVIEW_SCHEDULED",
              "HIRED",
              "REJECTED",
              "WITHDRAWN"
            ]
          },
          "appliedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 255,
            "minLength": 3
          },
          "password": {
            "type": "string",
            "maxLength": 255,
            "minLength": 8
          },
          "confirmPassword": {
            "type": "string",
            "maxLength": 255,
            "minLength": 8
          },
          "email": {
            "type": "string",
            "format": "email",
            "minLength": 1
          },
          "firstName": {
            "type": "string",
            "minLength": 1
          },
          "lastName": {
            "type": "string",
            "minLength": 1
          },
          "gender": {
            "type": "string",
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER",
              "UNSPECIFIED"
            ]
          },
          "role": {
            "type": "string",
            "enum": [
              "SEEKER",
              "RECRUITER"
            ]
          },
          "phoneNumber": {
            "type": "string",
            "maxLength": 30,
            "minLength": 8,
            "pattern": "^\\+?[0-9 ]+$"
          }
        },
        "required": [
          "confirmPassword",
          "email",
          "firstName",
          "lastName",
          "password",
          "role",
          "username"
        ]
      },
      "RegisterResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "gender": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "SEEKER",
              "RECRUITER"
            ]
          },
          "phoneNumber": {
            "type": "string"
          },
          "registrationSource": {
            "type": "string"
          }
        }
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "paged": {
            "type": "boolean"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "unpaged": {
            "type": "boolean"
          }
        }
      },
...
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          }
        }
      },
...
      "ApiResponsePagePublicJobResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PagePublicJobResponse"
          }
        }
      },
      "PagePublicJobResponse": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PublicJobResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "PublicJobResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "companyId": {
            "type": "integer",
            "format": "int64"
          },
          "companyName": {
            "type": "string"
          },
          "categoryId": {
            "type": "integer",
            "format": "int64"
          },
          "categoryName": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "jobType": {
            "type": "string"
          },
          "workMode": {
            "type": "string"
          },
          "salaryMin": {
            "type": "number"
          },
          "salaryMax": {
            "type": "number"
          },
          "experienceLevel": {
            "type": "string"
          },
          "publishedAt": {
            "type": "string",
            "format": "date-time"
          },
          "expiredAt": {
            "type": "string",
            "format": "date-time"
          },
          "sections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobPostSectionResponse"
            }
          },
          "skills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobPostSkillResponse"
            }
          }
        }
      },
      "ApiResponsePublicJobResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PublicJobResponse"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}